+++ config/application.rb
@@ -23,5 +23,7 @@ module SaisonmanagerApi
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = false
+
+    config.middleware.use Rack::Attack
   end
 end